Add: adjusted routes for web-chat, basig designe
Adjusted routes for web-chat and chat, separated to keep isolated both excersices.
Create the new layout for web-chat excersice views.layout.chat-main
Implemented the design of login page for web-chat.
Implemented the basic design of chat room.

Implement of behavior of login with google-sing-in and custom sing-in.
Logic of google-sign-in and google-log-out separeted in diferent js files.

Change name of template views.chat => views.socket